Check difference betweeen threaded and non threaded versions
