+++ Makefile
@@ -1,0 +1,5 @@
+run: 
+	./a.out
+
+compile:
+	g++ -std=c++11 *.cpp
