+++ thread.h
@@ -1,0 +1,9 @@
+#pragma once
+
+using namespace std;
+
+void f_count(int , mutex );
+
+
+void threadedVersion();
+
